@@ -1,37 +1,35 @@
 using Microsoft.CodeAnalysis;
 
-namespace SourceGenerator
+namespace SourceGenerator;
+
+[Generator]
+public sealed class HelloSourceGenerator : IIncrementalGenerator
 {
-    [Generator]
-    public class HelloSourceGenerator : ISourceGenerator
+    public void Initialize(IncrementalGeneratorInitializationContext context)
     {
-        public void Execute(GeneratorExecutionContext context)
-        {
-            // Find the main method
-            var mainMethod = context.Compilation.GetEntryPoint(context.CancellationToken);
+        var compilationIncrementalValue = context.CompilationProvider;
 
-            // Build up the source code
-            string source = $@"// <auto-generated/>
-using System;
-
-namespace {mainMethod.ContainingNamespace.ToDisplayString()}
-{{
-    public static partial class {mainMethod.ContainingType.Name}
-    {{
-        static partial void HelloFrom(string name) =>
-            Console.WriteLine($""Generator says: Hi from '{{name}}'"");
-    }}
-}}
-";
+        context.RegisterSourceOutput(
+            compilationIncrementalValue, 
+            static (context, compilation) =>
+        {            
+            // Get the entry point method
+            var mainMethod = compilation.GetEntryPoint(context.CancellationToken);
             var typeName = mainMethod.ContainingType.Name;
+
+            string source = $$"""
+                // Auto-generated code
+                namespace {{mainMethod.ContainingNamespace.ToDisplayString()}};
+                
+                public static partial class {{typeName}}
+                {
+                    static partial void HelloFrom(string name) =>
+                        Console.WriteLine($"Generator says: Hi from '{name}'");
+                }                
+                """;
 
             // Add the source code to the compilation
             context.AddSource($"{typeName}.g.cs", source);
-        }
-
-        public void Initialize(GeneratorInitializationContext context)
-        {
-            // No initialization required for this one
-        }
+        });
     }
 }
